Raw indicator calculus

Raw indicator calculus

Usage

indicator_raw(
  numerador,
  denominador,
  denominador_type = "pop",
  keep_raw_values = FALSE,
  treat_inf_values = TRUE,
  nome,
  agg,
  multi,
  decimals
)

Arguments

numerador

character. Spatial aggregation level. uf_res for UF of residence. uf_ocor for UF of occurrence. regsaude_res for regiao de saude of residence. regsaude_ocor for regiao de saúde of occurence. mun_res for municipality of residence. mun_ocor for municipality of ocurrence.

denominador

numeric. Year of death.

denominador_type

character. Type of denominator. pop or arbitrary.

keep_raw_values

logical. Keep numerator and denominator values on results. Defaults to FALSE.

treat_inf_values

logical. Treat positive infinite values caused by denominators equals to zero as NA. Defaults to TRUE.

nome

character. PCDaS API token. If not provided, the function will look for it on renvirom.

agg

character. Aggregation acronymin.

multi

integer. Multiplicator for indicator.

decimals

integer. Number of decimals for indicator.